clean but high-overhead i/o for homopolymer event.
NOTE:
there are 4-byte per 12-byte match event now, which is huge overhead
we can save 4 byte by storing 2byte hp-length and 2byte base length if needed
we can also save 4 bytes by writing homopolymer events to a different stream
this can be done down the line if we have time

+++ src/main/java/com/bina/lrsim/bioinfo/Context.java
@@ -11,9 +11,9 @@ import java.util.Iterator;
  */
 public class Context {
     private int kmer_;
-    private short hp_len_;
+    private int hp_len_;
 
-    public Context(int kmer, short hp_len) {
+    public Context(int kmer, int hp_len) {
         kmer_ = kmer;
         hp_len_ = hp_len;
     }
@@ -26,7 +26,7 @@ public class Context {
         return kmer_;
     }
 
-    public final short hp_len() {
+    public final int hp_len() {
         return hp_len_;
     }
